+++ Compiler/Compile/SPIRVLink.cs
@@ -50,7 +50,7 @@ namespace SSLang
 			{
 				proc.StartInfo = psi;
 				proc.Start();
-				bool done = proc.WaitForExit(options.CompilerTimeout);
+				bool done = proc.WaitForExit((int)options.CompilerTimeout);
 				if (!done)
 				{
 					proc.Kill();
