+++ Compiler/Compile/SPIRVOpt.cs
@@ -38,7 +38,7 @@ namespace SSLang
 			{
 				proc.StartInfo = psi;
 				proc.Start();
-				bool done = proc.WaitForExit(options.CompilerTimeout);
+				bool done = proc.WaitForExit((int)options.CompilerTimeout);
 				if (!done)
 				{
 					proc.Kill();
